fix(noise)+test: close adversarially-verified false-PASS floor and Java-wrap UB
28-에이전트 적대 검증 결과 반영:
- test_noise blended: n_vec >= 100 하한이 꼬리 절단/부분 삭제를 통과시키는
  false-PASS 채널임을 변조 실험으로 확인 — 동결 golden 의 상수 680 으로 고정
- pow2i 를 ldexp 로 교체: (n+1023)<<52 는 normal 범위 밖에서 조용히 틀리고
  int 오버플로 UB. ldexp 는 전체 int 지수 구간에서 Math.pow(2,n) 과 비트 동일
- octave 키 first_octave+i, legacy -first_octave / count-1, expectedDeviation
  n+1 을 Java iadd/ineg 래핑 시맨틱으로 계산 (C signed 오버플로 UB 제거) —
  6b 데이터팩 JSON 의 firstOctave 는 스키마상 무제한 int 라 방어가 실질

11/11 ctest, no-FMA, ASan+UBSan 게이트 전부 PASS 유지.

+++ core/src/noise_normal.c
@@ -8,9 +8,10 @@
 /* NormalNoise.INPUT_FACTOR — second 좌표 스케일 */
 #define HC_NORMAL_INPUT_FACTOR 1.0181268882175227
 
-/* expectedDeviation(n) = 0.1 * (1 + 1/(n+1)) — 결합 순서 그대로 */
+/* expectedDeviation(n) = 0.1 * (1 + 1/(n+1)) — 결합 순서 그대로.
+ * n+1 도 Java iadd 래핑으로 계산한다 (C signed 오버플로 UB 회피). */
 static double expected_deviation(int32_t n) {
-    return 0.1 * (1.0 + 1.0 / (double)(n + 1));
+    return 0.1 * (1.0 + 1.0 / (double)(int32_t)((uint32_t)n + 1u));
 }
 
 int hc_normal_noise_init(hc_normal_noise_t *n, hc_arena_t *a, hc_xoro_t *rand,
